organise components and update imports
